+++ Changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add download function to webnodejs engine. Completed on 2025-05-15
 - Apply optional `Content-Type` checking to download function at webnodejs engine. Completed on 2025-05-15
 - Add downlod flag to `atomic/atom/smfetch/src/browser/smfetch/atom.js` which return data as blob data type. Completed on 2025-05-15
+- Add `compmgr` engine which will manage components project. It will instead the `OricommJS_Components`. Completed on 2025-06-08
 
 ### Changed
 
@@ -32,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix session checking bypass the static file request in `webserver.js` at webnodejs engine. Completed on 2025-05-02
-- Fix mising checking reroute parameter in deskfetch method at `atomic/atom/smfetch/src/browser/atom.js` which will failure to automatically change the content of page. Completed on 2025-06-03
 
 ### Security